+++ packages/provider-queries/src/article.web.js
@@ -94,6 +94,8 @@ export default gql`
     publicationName
     publishedTime
     section
+    shortIdentifier
+    slug
     url
   }