+++ ios-app/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.3.36"></a>
+## [0.3.36](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.3.35...@times-components/ios-app@0.3.36) (2018-08-24)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+<a name="0.3.35"></a>
+## [0.3.35](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.3.34...@times-components/ios-app@0.3.35) (2018-08-24)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+<a name="0.3.34"></a>
+## [0.3.34](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.3.33...@times-components/ios-app@0.3.34) (2018-08-23)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+<a name="0.3.33"></a>
+## [0.3.33](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.3.32...@times-components/ios-app@0.3.33) (2018-08-23)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
 <a name="0.3.32"></a>
 ## [0.3.32](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.3.31...@times-components/ios-app@0.3.32) (2018-08-23)